refactor: updata new way about read yamlFile

+++ service/authService.go
@@ -44,7 +44,7 @@ func GetInfo(token string) (*model.User, error) {
 	user.Email = res["email"].(string)
 	user.Tel = res["phone"].(string)
 	user.QQ = res["qq"].(string)
-	user.Gender = int(res["gender"].(float64)) // interface无法直接转为int类型
+	user.Gender = res["gender"].(int)
 	user.IdcardNumber = res["id"].(string)
 	user.School = common.Whut
 	return user, nil
